# docs/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Brasa.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/menu/items/{itemId}": {
      "delete": {
        "tags": [
          "CatalogEndpoints"
        ],
        "summary": "Soft-deletes a menu item. Past orders keep their own snapshot of its name and price.",
        "operationId": "DeleteMenuItem",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/menu/items/{itemId}/details": {
      "put": {
        "tags": [
          "CatalogEndpoints"
        ],
        "summary": "Sets a menu item's description and declared allergens (CAT-02).",
        "operationId": "UpdateMenuItemDetails",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuItemDetailsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/menu/items/import": {
      "post": {
        "tags": [
          "CatalogEndpoints"
        ],
        "summary": "Bulk-creates menu items from a CSV file (CAT-17).",
        "operationId": "ImportMenuItems",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportMenuItemsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/menu/items/{itemId}/availability": {
      "put": {
        "tags": [
          "CatalogEndpoints"
        ],
        "summary": "86's a menu item, or brings it back (CAT-13).",
        "operationId": "UpdateMenuItemAvailability",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuItemAvailabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/menu/items/{itemId}/price": {
      "put": {
        "tags": [
          "CatalogEndpoints"
        ],
        "summary": "Changes a menu item's price for future orders. Past order lines keep their own snapshot.",
        "operationId": "UpdateMenuItemPrice",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuItemPriceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/menu/items/{itemId}/takeaway-price": {
      "put": {
        "tags": [
          "CatalogEndpoints"
        ],
        "summary": "Sets or clears a menu item's separate takeaway price (CAT-06). Null falls back to the dine-in price.",
        "operationId": "UpdateMenuItemTakeawayPrice",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuItemTakeawayPriceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/menu/items/{itemId}/course": {
      "put": {
        "tags": [
          "CatalogEndpoints"
        ],
        "summary": "Sets or clears which course a menu item is served at (CAT-14).",
        "operationId": "UpdateMenuItemCourse",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuItemCourseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/menu/items/{itemId}/station": {
      "put": {
        "tags": [
          "CatalogEndpoints"
        ],
        "summary": "Sets or clears which kitchen station prepares a menu item (CAT-15).",
        "operationId": "UpdateMenuItemStation",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuItemStationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/menu/categories/{categoryId}/visibility": {
      "put": {
        "tags": [
          "CatalogEndpoints"
        ],
        "summary": "Hides a whole category (and every item under it) from GET /menu, or shows it again (CAT-01).",
        "operationId": "UpdateMenuCategoryVisibility",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuCategoryVisibilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ImportMenuItemsRequest": {
        "required": [
          "csv"
        ],
        "type": "object",
        "properties": {
          "csv": {
            "type": "string"
          }
        }
      },
...
      "UpdateMenuCategoryVisibilityRequest": {
        "required": [
          "isVisible"
        ],
        "type": "object",
        "properties": {
          "isVisible": {
            "type": "boolean"
          }
        }
      },
      "UpdateMenuItemAvailabilityRequest": {
        "required": [
          "isAvailable"
        ],
        "type": "object",
        "properties": {
          "isAvailable": {
            "type": "boolean"
          }
        }
      },
      "UpdateMenuItemCourseRequest": {
        "required": [
          "course"
        ],
        "type": "object",
        "properties": {
          "course": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "UpdateMenuItemDetailsRequest": {
        "required": [
          "description",
          "allergens"
        ],
        "type": "object",
        "properties": {
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "allergens": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UpdateMenuItemPriceRequest": {
        "required": [
          "price"
        ],
        "type": "object",
        "properties": {
          "price": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "UpdateMenuItemStationRequest": {
        "required": [
          "station"
        ],
        "type": "object",
        "properties": {
          "station": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "UpdateMenuItemTakeawayPriceRequest": {
        "required": [
          "price"
        ],
        "type": "object",
        "properties": {
          "price": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
    }
  },
...
}